% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lizard_data}
\alias{lizard_data}
\title{Florida Scrub Lizard Data}
\format{
A data frame.
}
\source{
https://doi.org/10.5061/dryad.dbrv15f00
}
\usage{
lizard_data
}
\description{
Size, sex, reproductive status and body temperature dataset by Alison Gainsbury (2021).
}
\keyword{datasets}
