% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{two_sample_bootstrap}
\alias{two_sample_bootstrap}
\title{Two-Sample Bootstrap Method}
\usage{
two_sample_bootstrap(
  group1,
  group2,
  iterations = 1000,
  stat = "mean",
  resample_length = NULL
)
}
\arguments{
\item{group1}{A numeric vector representing the first biological sample (e.g., Male Snout Length).}

\item{group2}{A numeric vector representing the second biological sample (e.g., Female Snout Length).}

\item{iterations}{A numeric value specifying the number of bootstrap samples to generate. Default is 1000.}

\item{stat}{A character string specifying the statistic of interest: either "mean" or "median". Default is "mean".}

\item{resample_length}{A description of what this parameter does.}
}
\value{
A numeric vector containing the bootstrap distribution of the differences (\code{group1} - \code{group2}).
}
\description{
Calculates the bootstrap distribution of the difference between
two biological groups using either the difference in means or the difference in medians.
}
\examples{
set.seed(123)
# Simulate some fake lizard snout length data
male_lizards <- rnorm(30, mean = 45, sd = 5)
female_lizards <- rnorm(30, mean = 50, sd = 5)

# Run the bootstrap function for difference in means
boot_results <- two_sample_bootstrap(male_lizards, female_lizards, iterations = 1000, stat = "mean")

}
